+++ app/src/main/java/com/battlelancer/seriesguide/util/RatingsTools.kt
@@ -11,6 +11,7 @@ import com.battlelancer.seriesguide.R
 import com.battlelancer.seriesguide.databinding.LayoutRatingsBinding
 import com.battlelancer.seriesguide.shows.database.SgEpisode2
 import com.battlelancer.seriesguide.traktapi.TraktTools
+import com.battlelancer.seriesguide.ui.widgets.RatingView
 import com.uwetrottmann.androidutils.AndroidUtils
 import java.math.BigDecimal
 import java.math.RoundingMode
@@ -45,6 +46,19 @@ object RatingsTools {
         } else {
             groupRatingsUser.isGone = true
         }
+    }
+
+    fun RatingView.setLink(context: Context, url: String) {
+        isFocusable = true
+        setOnClickListener {
+            WebTools.openInApp(context, url)
+        }
+        setBackgroundResource(
+            ThemeUtils.resolveAttributeToResourceId(
+                context.theme,
+                android.R.attr.selectableItemBackground
+            )
+        )
     }
 
     fun LayoutRatingsBinding.setRatingValues(